Refactor register submit to actions.ts
